refactor tests into same package
