+++ src/org/jrm/FileOutputTest.java
@@ -1,4 +1,4 @@
-package org.jrm.test;
+package org.jrm;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
